requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
numpy==2.3.0
opencv-python-headless==4.11.0.86
pillow==11.2.1
scikit-learn==1.7.0
scipy==1.15.3
threadpoolctl==3.6.0
Werkzeug==3.1.3